+++ TeamCode/build.gradle
@@ -28,7 +28,7 @@ dependencies {
     annotationProcessor files('lib/OpModeAnnotationProcessor.jar')
 }
 dependencies {
-    implementation 'org.openftc:easyopencv:1.5.2'
+    implementation 'org.openftc:easyopencv:1.6.0'
 }
 
 repositories {
