+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Base/MainBase.java
@@ -1,8 +1,12 @@
 package org.firstinspires.ftc.teamcode.Base;
 
+import android.util.Log;
+
+import com.qualcomm.hardware.kauailabs.NavxMicroNavigationSensor;
 import com.qualcomm.hardware.modernrobotics.ModernRoboticsI2cGyro;
 import com.qualcomm.hardware.modernrobotics.ModernRoboticsI2cRangeSensor;
 import com.qualcomm.hardware.rev.Rev2mDistanceSensor;
+import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.eventloop.opmode.OpMode;
 import com.qualcomm.robotcore.hardware.ColorSensor;
@@ -14,6 +18,11 @@ import com.qualcomm.robotcore.util.Range;
 
 import org.firstinspires.ftc.robotcore.external.navigation.Acceleration;
 import org.firstinspires.ftc.robotcore.external.navigation.Orientation;
+import org.firstinspires.ftc.teamcode.android.navx_ftc.src.main.java.com.kauailabs.navx.ftc.AHRS;
+import org.firstinspires.ftc.teamcode.android.navx_ftc.src.main.java.com.kauailabs.navx.ftc.navXPIDController;
+
+import java.text.DecimalFormat;
+
 
 public class MainBase {
 
@@ -30,11 +39,11 @@ public class MainBase {
     public DcMotor rightgrabber = null;
     public DcMotor leftgrabber = null;
 
-
+    // Total Sensors: 4
     public ColorSensor color = null;
-//    ModernRoboticsI2cRangeSensor rangebrothers;
-//    Rev2mDistanceSensor rangebrothers2;
     public ModernRoboticsI2cGyro gyro = null;
+    public AHRS navx_device;
+    public navXPIDController yawPIDController;
 
 
 
@@ -73,7 +82,27 @@ public class MainBase {
 
 
 
-    private ElapsedTime runtime = new ElapsedTime();
+    public ElapsedTime runtime = new ElapsedTime();
+
+
+    // variables for navx
+    public final byte NAVX_DEVICE_UPDATE_RATE_HZ = 50;
+
+    public final double TARGET_ANGLE_DEGREES = 0.0;
+    public final double TOLERANCE_DEGREES = 2.0;
+    public final double MIN_MOTOR_OUTPUT_VALUE = -1.0;
+    public final double MAX_MOTOR_OUTPUT_VALUE = 1.0;
+    public final double YAW_PID_P = 0.001;
+    public final double YAW_PID_I = 0.0003;
+    public final double YAW_PID_D = 0.0003;
+
+
+    public boolean calibration_complete = false;
+
+
+    public double limit(double a) {
+        return Math.min(Math.max(a, MIN_MOTOR_OUTPUT_VALUE), MAX_MOTOR_OUTPUT_VALUE);
+    }
 
     /**
      *  initialize
@@ -96,6 +125,23 @@ public class MainBase {
 
         gyro = hwMap.get(ModernRoboticsI2cGyro.class,"Gyro");
         color = hwMap.get(ColorSensor.class,"sensor_color");
+        navx_device = AHRS.getInstance(hwMap.get(NavxMicroNavigationSensor.class, "navx"),
+                AHRS.DeviceDataType.kProcessedData,
+                NAVX_DEVICE_UPDATE_RATE_HZ);
+
+
+        /* Create a PID Controller which uses the Yaw Angle as input. */
+        yawPIDController = new navXPIDController( navx_device,
+                navXPIDController.navXTimestampedDataSource.YAW);
+
+        /* Configure the PID controller */
+        yawPIDController.setSetpoint(TARGET_ANGLE_DEGREES);
+        yawPIDController.setContinuous(true);
+        yawPIDController.setOutputRange(MIN_MOTOR_OUTPUT_VALUE, MAX_MOTOR_OUTPUT_VALUE);
+        yawPIDController.setTolerance(navXPIDController.ToleranceType.ABSOLUTE, TOLERANCE_DEGREES);
+        yawPIDController.setPID(YAW_PID_P, YAW_PID_I, YAW_PID_D);
+        yawPIDController.enable(true);
+
 
 
         backleft.setDirection(DcMotor.Direction.FORWARD);
@@ -105,7 +151,6 @@ public class MainBase {
         lift.setDirection(DcMotor.Direction.FORWARD);
         leftgrabber.setDirection(DcMotor.Direction.FORWARD);
         rightgrabber.setDirection(DcMotor.Direction.FORWARD);
-
 
 
         backleft.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
@@ -152,13 +197,442 @@ public class MainBase {
         opMode.telemetry.update();
     }
 
+
     // ***************************************************************
     // Public functions
     // ***************************************************************
 
 
 
-    public void gyroStrafe(double inches,double angle, double speed, LinearOpMode opMode) {
+
+
+
+
+
+
+
+//    public void navxgyroTurn(double speed, double angle , LinearOpMode opmode) {
+//        frontleft.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
+//        backleft.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
+//        frontright.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
+//        backright.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
+//        yawPIDController.setSetpoint(angle);
+//        // keep looping while we are still active, and not on heading.
+//        while (opmode.opModeIsActive() && yawPIDController.getError() <= HEADING_THRESHOLD) {
+//            // Update telemetry & Allow time for other processes to run.
+//            opMode.telemetry.update();
+//            frontleft.setPower(-speed);
+//            backleft.setPower(-speed);
+//            frontright.setPower(speed);
+//            backright.setPower(speed);
+//        }
+//        frontleft.setMode(DcMotor.RunMode.RUN_TO_POSITION);
+//        backleft.setMode(DcMotor.RunMode.RUN_TO_POSITION);
+//        frontright.setMode(DcMotor.RunMode.RUN_TO_POSITION);
+//        backright.setMode(DcMotor.RunMode.RUN_TO_POSITION);
+
+
+
+
+
+    /**
+     * --------------------DOES NOT WORK-------------------------------
+     * this is the ConceptNavxDriveStraightPIDLinearOp from navx example made into a function
+     * used PID which is set above
+     * @param targetDistance
+     * @param opMode
+     */
+    public void navxDriveExample(double targetDistance, LinearOpMode opMode)  {
+
+        /* Create a PID Controller which uses the Yaw Angle as input. */
+        yawPIDController = new navXPIDController( navx_device,
+                navXPIDController.navXTimestampedDataSource.YAW);
+
+        /* Configure the PID controller */
+        yawPIDController.setSetpoint(targetDistance);
+        yawPIDController.setContinuous(true);
+        yawPIDController.setOutputRange(MIN_MOTOR_OUTPUT_VALUE, MAX_MOTOR_OUTPUT_VALUE);
+        yawPIDController.setTolerance(navXPIDController.ToleranceType.ABSOLUTE, TOLERANCE_DEGREES);
+        yawPIDController.setPID(YAW_PID_P, YAW_PID_I, YAW_PID_D);
+        yawPIDController.enable(true);
+
+        while ( !calibration_complete ) {
+            /* navX-Micro Calibration completes automatically ~15 seconds after it is
+            powered on, as long as the device is still.  To handle the case where the
+            navX-Micro has not been able to calibrate successfully, hold off using
+            the navX-Micro Yaw value until calibration is complete.
+             */
+            calibration_complete = !navx_device.isCalibrating();
+            if (!calibration_complete) {
+                opMode.telemetry.addData("navX-Micro", "Startup Calibration in Progress");
+            }
+        }
+        navx_device.zeroYaw();
+
+        /* Wait for new Yaw PID output values, then update the motors
+           with the new PID value with each new output value.
+         */
+
+        final double TOTAL_RUN_TIME_SECONDS = 10.0;
+        int DEVICE_TIMEOUT_MS = 500;
+        navXPIDController.PIDResult yawPIDResult = new navXPIDController.PIDResult();
+
+        /* Drive straight forward at 1/2 of full drive speed */
+        double drive_speed = 0.5;
+
+        DecimalFormat df = new DecimalFormat("#.##");
+
+        try {
+            while ((runtime.time() < TOTAL_RUN_TIME_SECONDS) &&
+                    !Thread.currentThread().isInterrupted()) {
+                if (yawPIDController.waitForNewUpdate(yawPIDResult, DEVICE_TIMEOUT_MS)) {
+                    if (yawPIDResult.isOnTarget()) {
+                        frontleft.setPower(drive_speed);
+                        frontright.setPower(drive_speed);
+                        backleft.setPower(drive_speed);
+                        backright.setPower(drive_speed);
+                        opMode.telemetry.addData("PIDOutput", df.format(drive_speed) + ", " +
+                                df.format(drive_speed));
+                    } else {
+                        double output = yawPIDResult.getOutput();
+                        frontright.setPower(drive_speed + output);
+                        frontleft.setPower(drive_speed - output);
+                        backleft.setPower(drive_speed - output);
+                        backright.setPower(drive_speed + output);
+                        opMode.telemetry.addData("PIDOutput", df.format(limit(drive_speed + output)) + ", " +
+                                df.format(limit(drive_speed - output)));
+                    }
+                    opMode.telemetry.addData("Yaw", df.format(navx_device.getYaw()));
+                } else{
+                    /* A timeout occurred */
+                    Log.w("navXDriveStraightOp", "Yaw PID waitForNewUpdate() TIMEOUT.");
+                }
+            }
+        }
+        catch(InterruptedException ex) {
+            Thread.currentThread().interrupt();
+        }
+        finally {
+            navx_device.close();
+            opMode.telemetry.addData("LinearOp", "Complete");
+        }
+    }
+
+
+    /**
+     * --------------------------WORKING ZONE----------------------------
+     *
+     * @param speed
+     * @param frontLeftInches
+     * @param frontRightInches
+     * @param backLeftInches
+     * @param backRightInches
+     * @param angle
+     * @param opMode
+     */
+    public void navxgyroStrafe(double speed,
+                              double frontLeftInches, double frontRightInches, double backLeftInches,
+                              double backRightInches,
+                              double angle, LinearOpMode opMode) {
+
+        int newFrontLeftTarget;
+        int newFrontRightTarget;
+        int newBackLeftTarget;
+        int newBackRightTarget;
+
+
+
+        yawPIDController = new navXPIDController( navx_device,
+                navXPIDController.navXTimestampedDataSource.YAW);
+        //AHRS navx_device;
+        navXPIDController.PIDResult yawPIDResult = new navXPIDController.PIDResult();
+
+        yawPIDController.setSetpoint(angle);
+
+        opMode.telemetry.addData("yawPIDController.getSetpoint()",yawPIDController.getSetpoint());
+        opMode.telemetry.update();
+
+        double HalfMaxOne;
+        double HalfMaxTwo;
+
+        double max;
+
+        double error;
+        double steer;
+        double frontLeftSpeed;
+        double frontRightSpeed;
+        double backLeftSpeed;
+        double backRightSpeed;
+
+        double ErrorAmount;
+        boolean goodEnough = false;
+
+        // Ensure that the opmode is still active
+        if (opMode.opModeIsActive()) {
+
+            opMode.telemetry.addData("yaw", navx_device.getYaw());
+            opMode.telemetry.update();
+
+            // Determine new target position, and pass to motor controller
+            newFrontLeftTarget = frontleft.getCurrentPosition() + (int) (frontLeftInches * cpi);
+            newFrontRightTarget = frontright.getCurrentPosition() - (int) (frontRightInches * cpi);
+            newBackLeftTarget = backleft.getCurrentPosition() - (int) (backLeftInches * cpi);
+            newBackRightTarget = backright.getCurrentPosition() + (int) (backRightInches * cpi);
+
+
+            // Set Target and Turn On RUN_TO_POSITION
+            frontleft.setTargetPosition(newFrontLeftTarget);
+            frontright.setTargetPosition(newFrontRightTarget);
+            backleft.setTargetPosition(newBackLeftTarget);
+            backright.setTargetPosition(newBackRightTarget);
+
+            frontleft.setMode(DcMotor.RunMode.RUN_TO_POSITION);
+            frontright.setMode(DcMotor.RunMode.RUN_TO_POSITION);
+            backleft.setMode(DcMotor.RunMode.RUN_TO_POSITION);
+            backright.setMode(DcMotor.RunMode.RUN_TO_POSITION);
+
+            frontleft.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
+            backright.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
+            backleft.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
+            frontright.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
+
+
+
+            // start motion.
+            speed = Range.clip(Math.abs(speed), 0.0, 1.0);
+            frontleft.setPower(Math.abs(speed));
+            frontright.setPower(Math.abs(speed));
+            backleft.setPower(Math.abs(speed));
+            backright.setPower(Math.abs(speed));
+            // keep looping while we are still active, and BOTH motors are running.
+            while (opMode.opModeIsActive()   // run as long as the operation mode is active. e.g. stop when ...
+                    && ((frontleft.isBusy() && frontright.isBusy()) && (backleft.isBusy() && backright.isBusy()))
+                    && !goodEnough) {
+
+                opMode.telemetry.addData("yawPIDController.getSetpoint()",yawPIDController.getSetpoint());
+                opMode.telemetry.update();
+                // adjust relative speed based on heading error.
+                steer = yawPIDResult.getOutput();
+
+                // if driving in reverse, the motor correction also needs to be reversed
+                if (frontLeftInches < 0 && frontRightInches < 0 && backLeftInches < 0 && backRightInches < 0)
+                    steer *= -1.0;
+
+                frontLeftSpeed = speed - steer;
+                backLeftSpeed = speed + steer;
+                backRightSpeed = speed + steer;
+                frontRightSpeed = speed - steer;
+
+                // Normalize speeds if either one exceeds +/- 1.0;
+                HalfMaxOne = Math.max(Math.abs(frontLeftSpeed), Math.abs(backLeftSpeed));
+                HalfMaxTwo = Math.max(Math.abs(frontRightSpeed), Math.abs(backRightSpeed));
+                max = Math.max(Math.abs(HalfMaxOne), Math.abs(HalfMaxTwo));
+                if (max > 1.0) {
+                    frontLeftSpeed /= max;
+                    frontRightSpeed /= max;
+                    backLeftSpeed /= max;
+                    backRightSpeed /= max;
+                }
+
+                frontleft.setPower(frontLeftSpeed);
+                frontright.setPower(frontRightSpeed);
+                backleft.setPower(backLeftSpeed);
+                backright.setPower(backRightSpeed);
+
+                // Display drive status for the driver.
+//                telemetry.addData("Err/St", "%5.1f/%5.1f", error, steer);
+//                telemetry.addData("Target", "%7d:%7d", newBackLeftTarget, newBackRightTarget, newFrontLeftTarget, newFrontRightTarget);
+//                telemetry.addData("Actual", "%7d:%7d", backleft.getCurrentPosition(), backright.getCurrentPosition(), frontleft.getCurrentPosition(), frontright.getCurrentPosition());
+//                telemetry.addData("Speed", "%5.2f:%5.2f", backLeftSpeed, backRightSpeed, frontLeftSpeed, frontRightSpeed);
+//                telemetry.update();
+
+                ErrorAmount = ((Math.abs(((newBackLeftTarget) - (backleft.getCurrentPosition())))
+                        + (Math.abs(((newFrontLeftTarget) - (frontleft.getCurrentPosition()))))
+                        + (Math.abs((newBackRightTarget) - (backright.getCurrentPosition())))
+                        + (Math.abs(((newFrontRightTarget) - (frontright.getCurrentPosition()))))) / cpi);
+                if (ErrorAmount < amountError) {
+                    goodEnough = true;
+                }
+            }
+
+            // Stop all motion;
+            frontleft.setPower(0);
+            frontright.setPower(0);
+            backleft.setPower(0);
+            backright.setPower(0);
+
+            // Turn off RUN_TO_POSITION
+
+            frontleft.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
+            frontright.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
+            backleft.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
+            backright.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
+        }
+    }
+
+
+
+
+    /**
+     *
+     * Drives using navx gyro. References Yahya's code
+     * @param speed
+     * @param frontLeftInches
+     * @param frontRightInches
+     * @param backLeftInches
+     * @param backRightInches
+     * @param angle
+     * @param opMode
+     */
+    public void navxgyroDrive(double speed,
+                          double frontLeftInches, double frontRightInches, double backLeftInches,
+                          double backRightInches,
+                          double angle, LinearOpMode opMode) {
+
+        int newFrontLeftTarget;
+        int newFrontRightTarget;
+        int newBackLeftTarget;
+        int newBackRightTarget;
+
+
+
+        yawPIDController = new navXPIDController( navx_device,
+                navXPIDController.navXTimestampedDataSource.YAW);
+        //AHRS navx_device;
+        navXPIDController.PIDResult yawPIDResult = new navXPIDController.PIDResult();
+
+        yawPIDController.setSetpoint(angle);
+
+        opMode.telemetry.addData("yawPIDController.getSetpoint()",yawPIDController.getSetpoint());
+        opMode.telemetry.update();
+
+        double HalfMaxOne;
+        double HalfMaxTwo;
+
+        double max;
+
+        double error;
+        double steer;
+        double frontLeftSpeed;
+        double frontRightSpeed;
+        double backLeftSpeed;
+        double backRightSpeed;
+
+        double ErrorAmount;
+        boolean goodEnough = false;
+
+        // Ensure that the opmode is still active
+        if (opMode.opModeIsActive()) {
+
+            opMode.telemetry.addData("yawPIDController.getSetpoint()",yawPIDController.getSetpoint());
+            opMode.telemetry.update();
+
+            // Determine new target position, and pass to motor controller
+            newFrontLeftTarget = frontleft.getCurrentPosition() + (int) (frontLeftInches * cpi);
+            newFrontRightTarget = frontright.getCurrentPosition() + (int) (frontRightInches * cpi);
+            newBackLeftTarget = backleft.getCurrentPosition() + (int) (backLeftInches * cpi);
+            newBackRightTarget = backright.getCurrentPosition() + (int) (backRightInches * cpi);
+
+
+            // Set Target and Turn On RUN_TO_POSITION
+            frontleft.setTargetPosition(newFrontLeftTarget);
+            frontright.setTargetPosition(newFrontRightTarget);
+            backleft.setTargetPosition(newBackLeftTarget);
+            backright.setTargetPosition(newBackRightTarget);
+
+            frontleft.setMode(DcMotor.RunMode.RUN_TO_POSITION);
+            frontright.setMode(DcMotor.RunMode.RUN_TO_POSITION);
+            backleft.setMode(DcMotor.RunMode.RUN_TO_POSITION);
+            backright.setMode(DcMotor.RunMode.RUN_TO_POSITION);
+
+            frontleft.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
+            backright.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
+            backleft.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
+            frontright.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
+
+
+
+            // start motion.
+            speed = Range.clip(Math.abs(speed), 0.0, 1.0);
+            frontleft.setPower(Math.abs(speed));
+            frontright.setPower(Math.abs(speed));
+            backleft.setPower(Math.abs(speed));
+            backright.setPower(Math.abs(speed));
+            // keep looping while we are still active, and BOTH motors are running.
+            while (opMode.opModeIsActive()   // run as long as the operation mode is active. e.g. stop when ...
+                    && ((frontleft.isBusy() && frontright.isBusy()) && (backleft.isBusy() && backright.isBusy()))
+                    && !goodEnough) {
+
+                opMode.telemetry.addData("yawPIDController.getSetpoint()",yawPIDController.getSetpoint());
+                opMode.telemetry.update();
+                // adjust relative speed based on heading error.
+                steer = yawPIDResult.getOutput();
+
+                // if driving in reverse, the motor correction also needs to be reversed
+                if (frontLeftInches < 0 && frontRightInches < 0 && backLeftInches < 0 && backRightInches < 0)
+                    steer *= -1.0;
+
+                frontLeftSpeed = speed - steer;
+                backLeftSpeed = speed - steer;
+                backRightSpeed = speed + steer;
+                frontRightSpeed = speed + steer;
+
+                // Normalize speeds if either one exceeds +/- 1.0;
+                HalfMaxOne = Math.max(Math.abs(frontLeftSpeed), Math.abs(backLeftSpeed));
+                HalfMaxTwo = Math.max(Math.abs(frontRightSpeed), Math.abs(backRightSpeed));
+                max = Math.max(Math.abs(HalfMaxOne), Math.abs(HalfMaxTwo));
+                if (max > 1.0) {
+                    frontLeftSpeed /= max;
+                    frontRightSpeed /= max;
+                    backLeftSpeed /= max;
+                    backRightSpeed /= max;
+                }
+
+                frontleft.setPower(frontLeftSpeed);
+                frontright.setPower(frontRightSpeed);
+                backleft.setPower(backLeftSpeed);
+                backright.setPower(backRightSpeed);
+
+                // Display drive status for the driver.
+//                telemetry.addData("Err/St", "%5.1f/%5.1f", error, steer);
+//                telemetry.addData("Target", "%7d:%7d", newBackLeftTarget, newBackRightTarget, newFrontLeftTarget, newFrontRightTarget);
+//                telemetry.addData("Actual", "%7d:%7d", backleft.getCurrentPosition(), backright.getCurrentPosition(), frontleft.getCurrentPosition(), frontright.getCurrentPosition());
+//                telemetry.addData("Speed", "%5.2f:%5.2f", backLeftSpeed, backRightSpeed, frontLeftSpeed, frontRightSpeed);
+//                telemetry.update();
+
+                ErrorAmount = ((Math.abs(((newBackLeftTarget) - (backleft.getCurrentPosition())))
+                        + (Math.abs(((newFrontLeftTarget) - (frontleft.getCurrentPosition()))))
+                        + (Math.abs((newBackRightTarget) - (backright.getCurrentPosition())))
+                        + (Math.abs(((newFrontRightTarget) - (frontright.getCurrentPosition()))))) / cpi);
+                if (ErrorAmount < amountError) {
+                    goodEnough = true;
+                }
+            }
+
+            // Stop all motion;
+            frontleft.setPower(0);
+            frontright.setPower(0);
+            backleft.setPower(0);
+            backright.setPower(0);
+
+            // Turn off RUN_TO_POSITION
+
+            frontleft.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
+            frontright.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
+            backleft.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
+            backright.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
+        }
+    }
+
+
+    /**
+     * Strafing using modern robotics gyro
+     * @param inches
+     * @param angle
+     * @param speed
+     * @param opMode
+     */
+    public void MRgyroStrafe(double inches,double angle, double speed, LinearOpMode opMode) {
 //
         int move = (int) (Math.round(inches * cpi * meccyBias));
 //\
@@ -249,8 +723,11 @@ public class MainBase {
         backright.setPower(0);
         backleft.setPower(0);
     }
+
+
     /**
      * drives towards the backleft and frontright of the robot
+     * no gyro
      * @param inches
      * @param speed
      */
@@ -277,8 +754,11 @@ public class MainBase {
 
     }
 
+
+
     /**
      * drives towards the backright and frontleft of the robot
+     * no gyro
      * @param inches
      * @param speed
      */
@@ -308,6 +788,7 @@ public class MainBase {
     }
 
 
+
     /**
      * Find tape on field using color sensor
      * Change FIND_LINE_TIME to tell it how long it has to find the line before exiting program
@@ -334,6 +815,7 @@ public class MainBase {
     }
 
 
+
     /**
      * Lift operations
      * @param inches up is +, down is -
@@ -352,8 +834,10 @@ public class MainBase {
         lift.setPower(speed);
     }
 
+
+
     /**
-     * Grab cone
+     * Sucks cone
      */
     public void getItGirl(LinearOpMode opMode){
         while(lift.isBusy()){
@@ -366,6 +850,8 @@ public class MainBase {
         rightgrabber.setPower(0);
         leftgrabber.setPower(0);
     }
+
+
 
     /**
      * Let go of cone
@@ -383,6 +869,7 @@ public class MainBase {
         leftgrabber.setPower(0);
         goLift(6,.9);
     }
+
 
 
     /**
@@ -449,8 +936,9 @@ public class MainBase {
             backleft.setPower(Math.abs(speed));
             backright.setPower(Math.abs(speed));
             // keep looping while we are still active, and BOTH motors are running.
-            while (opMode.opModeIsActive() &&
-                    ((frontleft.isBusy() && frontright.isBusy()) && (backleft.isBusy() && backright.isBusy())) && !goodEnough) {
+            while (opMode.opModeIsActive()
+                    && ((frontleft.isBusy() && frontright.isBusy()) && (backleft.isBusy() && backright.isBusy()))
+                    && !goodEnough) {
 
 
                 // adjust relative speed based on heading error.
@@ -513,20 +1001,21 @@ public class MainBase {
         }
     }
 
+
+
     /**
      * Turning using Modern Robotics Gyro
      *
      * @param speed
      * @param angle
      */
-
     public void MRgyroTurn(double speed, double angle , LinearOpMode opmode) {
         frontleft.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
         backleft.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
         frontright.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
         backright.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
         // keep looping while we are still active, and not on heading.
-        while (opmode.opModeIsActive() && !onHeading(speed, angle, P_TURN_COEFF)) {
+        while (opmode.opModeIsActive() && !MRonHeading(speed, angle, P_TURN_COEFF)) {
             // Update telemetry & Allow time for other processes to run.
             opMode.telemetry.update();
         }
@@ -536,6 +1025,8 @@ public class MainBase {
         backright.setMode(DcMotor.RunMode.RUN_TO_POSITION);
 
     }
+
+
 
     /**
      * Using Modern Robotics Gyro
@@ -548,8 +1039,6 @@ public class MainBase {
      *                 If a relative angle is required, add/subtract from current heading.
      * @param holdTime Length of time (in seconds) to hold the specified heading.
      */
-
-
     public void gyroHold(double speed, double angle, double holdTime, LinearOpMode opMode) {
 
         ElapsedTime holdTimer = new ElapsedTime();
@@ -558,7 +1047,7 @@ public class MainBase {
         holdTimer.reset();
         while (opMode.opModeIsActive() && (holdTimer.time() < holdTime)) {
             // Update telemetry & Allow time for other processes to run.
-            onHeading(speed, angle, P_TURN_COEFF);
+            MRonHeading(speed, angle, P_TURN_COEFF);
             opMode.telemetry.update();
 
 
@@ -572,6 +1061,15 @@ public class MainBase {
     }
 
 
+
+
+    // ***************************************************************
+    // Private functions
+    // ***************************************************************
+
+
+
+
     /**
      * Using Modern Robotics Gyro
      * Perform one cycle of closed loop heading control.
@@ -583,7 +1081,7 @@ public class MainBase {
      * @param PCoeff Proportional Gain coefficient
      * @return
      */
-    boolean onHeading(double speed, double angle, double PCoeff) {
+    boolean MRonHeading(double speed, double angle, double PCoeff) {
         double error;
         double steer;
         boolean onTarget = false;
@@ -618,6 +1116,45 @@ public class MainBase {
         return onTarget;
     }
 
+    boolean navxonHeading(double speed, double angle, double PCoeff) {
+        double error;
+        double steer;
+        boolean onTarget = false;
+        double leftSpeed;
+        double rightSpeed;
+
+        yawPIDController.setSetpoint(angle);
+        // determine turn power based on +/- error
+        error = yawPIDController.getError();
+
+        if (Math.abs(error) <= HEADING_THRESHOLD) {
+            steer = 0.0;
+            leftSpeed = 0.0;
+            rightSpeed = 0.0;
+            onTarget = true;
+        } else {
+            steer = getSteer(error, PCoeff);
+            rightSpeed = speed * steer;
+            leftSpeed = -rightSpeed;
+        }
+
+        // Send desired speeds to motors.
+        frontleft.setPower(leftSpeed);
+        backleft.setPower(leftSpeed);
+        backright.setPower(rightSpeed);
+        frontright.setPower(rightSpeed);
+
+        // Display it for the driver.
+//        telemetry.addData("Target", "%5.2f", angle);
+//        telemetry.addData("Err/St", "%5.2f/%5.2f", error, steer);
+//        telemetry.addData("Speed.", "%5.2f:%5.2f", leftSpeed, rightSpeed);
+
+        return onTarget;
+    }
+
+
+
+
     /**
      * Using Modern Robotics Gyro
      * getError determines the error between the target angle and the robot's current heading
@@ -637,6 +1174,9 @@ public class MainBase {
         return -robotError;
     }
 
+
+
+
     /**
      * Using Modern Robotics Gyro
      * returns desired steering force.  +/- 1 range.  +ve = steer left
@@ -648,8 +1188,4 @@ public class MainBase {
     public double getSteer(double error, double PCoeff) {
         return Range.clip(error * PCoeff, -DRIVE_SPEED07, 1);
     }
-    // ***************************************************************
-    // Private functions
-    // ***************************************************************
-
 }
