+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Old/HugBot.java
@@ -1,5 +1,6 @@
-package org.firstinspires.ftc.teamcode.Test;
+package org.firstinspires.ftc.teamcode.Old;
 
+import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.OpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 import com.qualcomm.robotcore.hardware.DcMotor;
@@ -7,6 +8,7 @@ import com.qualcomm.robotcore.hardware.Servo;
 
 
 @TeleOp(name="HugBot")
+@Disabled
 public class HugBot extends OpMode {
