+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Old/TeleOP.java
@@ -1,11 +1,13 @@
-package org.firstinspires.ftc.teamcode.Test;
+package org.firstinspires.ftc.teamcode.Old;
 
+import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.OpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 import com.qualcomm.robotcore.hardware.DcMotor;
 
 
 @TeleOp(name="TeleOP")
+@Disabled
 public class TeleOP extends OpMode {
