+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Old/TeleOPColor.java
@@ -1,11 +1,13 @@
-package org.firstinspires.ftc.teamcode.Test;
+package org.firstinspires.ftc.teamcode.Old;
 
+import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.OpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 import com.qualcomm.robotcore.hardware.ColorSensor;
 
 
 @TeleOp(name="ColorTeleOP", group = "Sensor")
+@Disabled
 public class TeleOPColor extends OpMode {
 
     ColorSensor color;
