+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Test/PID.java
@@ -28,18 +28,20 @@ SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
 CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
 OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE. */
-package org.firstinspires.ftc.teamcode.android.OpModes;
+package org.firstinspires.ftc.teamcode.Test;
 
 import android.util.Log;
 
+import androidx.core.widget.TextViewCompat;
+
 import com.qualcomm.hardware.kauailabs.NavxMicroNavigationSensor;
 
-
+import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 import com.qualcomm.robotcore.hardware.DcMotor;
-import com.qualcomm.robotcore.hardware.DcMotorController;
 import com.qualcomm.robotcore.util.ElapsedTime;
+import com.qualcomm.robotcore.util.Range;
 
 import org.firstinspires.ftc.teamcode.android.navx_ftc.src.main.java.com.kauailabs.navx.ftc.AHRS;
 import org.firstinspires.ftc.teamcode.android.navx_ftc.src.main.java.com.kauailabs.navx.ftc.navXPIDController;
@@ -47,58 +49,83 @@ import org.firstinspires.ftc.teamcode.android.navx_ftc.src.main.java.com.kauaila
 import java.text.DecimalFormat;
 
 /*
- * An example linear op mode where the robot will rotate
- * to a specified angle an then stop.
+ * An example linear op mode where the robot will drive in
+ * a straight line (where the driving direction is guided by
+ * the Yaw angle from a navX-Model device).
  *
  * This example uses a simple PID controller configuration
  * with a P coefficient, and will likely need tuning in order
  * to achieve optimal performance.
  *
  * Note that for the best accuracy, a reasonably high update rate
- * for the navX-Model sensor should be used.
+ * for the navX-Model sensor should be used.  This example uses
+ * the default update rate (50Hz), which may be lowered in order
+ * to reduce the frequency of the updates to the drive system.
  */
-@TeleOp(name = "Concept: navX Rotate to Angle PID - Linear", group = "Concept")
+
+@Autonomous(name = "PID")
 // @Disabled Comment this in to remove this from the Driver Station OpMode List
-public class ConceptNavXRotateToAnglePIDLinearOp extends LinearOpMode {
-    DcMotor leftMotor;
-    DcMotor rightMotor;
+public class PID extends LinearOpMode {
+
+    DcMotor frontleft;
+    DcMotor frontright;
+    DcMotor backleft;
+    DcMotor backright;
+
+    double cpr = 11.5; //counts per rotation
+    Integer gearratio = 40;
+    Double diameter = 4.125;
+    Double cpi = (cpr * gearratio) / (Math.PI * diameter); //counts per inch, 28cpr * gear ratio / (2 * pi * diameter (in inches, in the center))
+    Double bias = 1.0;//default 0.8
+    Double meccyBias = 0.9;//change to adjust only strafing movement
 
     private AHRS navx_device;
-
     private navXPIDController yawPIDController;
     private ElapsedTime runtime = new ElapsedTime();
 
     private final byte NAVX_DEVICE_UPDATE_RATE_HZ = 50;
 
-    private final double TARGET_ANGLE_DEGREES = 90.0;
+    private final double TARGET_ANGLE_DEGREES = 0.0;
     private final double TOLERANCE_DEGREES = 2.0;
     private final double MIN_MOTOR_OUTPUT_VALUE = -1.0;
     private final double MAX_MOTOR_OUTPUT_VALUE = 1.0;
-    private final double YAW_PID_P = 0.005;
-    private final double YAW_PID_I = 0.0;
-    private final double YAW_PID_D = 0.0;
+    private final double YAW_PID_P = 0.00013;
+    private final double YAW_PID_I = 0.00013;
+    private final double YAW_PID_D = 0.0001;
+
+    static final double P_TURN_COEFF = 0.1;     // Larger is more responsive, but also less stable
 
     private boolean calibration_complete = false;
 
+    double amountError = 2;
+
+    public double limit(double a) {
+        return Math.min(Math.max(a, MIN_MOTOR_OUTPUT_VALUE), MAX_MOTOR_OUTPUT_VALUE);
+    }
+
     @Override
     public void runOpMode() throws InterruptedException {
-        leftMotor = hardwareMap.dcMotor.get("left motor");
-        rightMotor = hardwareMap.dcMotor.get("right motor");
+        frontleft = hardwareMap.dcMotor.get("frontleft");
+        frontright = hardwareMap.dcMotor.get("frontright");
+        backleft = hardwareMap.dcMotor.get("backleft");
+        backright = hardwareMap.dcMotor.get("backright");
 
         navx_device = AHRS.getInstance(hardwareMap.get(NavxMicroNavigationSensor.class, "navx"),
                 AHRS.DeviceDataType.kProcessedData,
                 NAVX_DEVICE_UPDATE_RATE_HZ);
 
-        rightMotor.setDirection(DcMotor.Direction.REVERSE);
+        frontright.setDirection(DcMotor.Direction.REVERSE);
+        backright.setDirection(DcMotor.Direction.REVERSE);
 
-        /* If possible, use encoders when driving, as it results in more */
-        /* predictable drive system response.                           */
-        //leftMotor.setChannelMode(DcMotorController.RunMode.RUN_USING_ENCODERS);
-        //rightMotor.setChannelMode(DcMotorController.RunMode.RUN_USING_ENCODERS);
+
+        frontright.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
+        frontleft.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
+        backleft.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
+        backright.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
 
         /* Create a PID Controller which uses the Yaw Angle as input. */
-        yawPIDController = new navXPIDController( navx_device,
-                                    navXPIDController.navXTimestampedDataSource.YAW);
+        yawPIDController = new navXPIDController(navx_device,
+                navXPIDController.navXTimestampedDataSource.YAW);
 
         /* Configure the PID controller */
         yawPIDController.setSetpoint(TARGET_ANGLE_DEGREES);
@@ -106,10 +133,9 @@ public class ConceptNavXRotateToAnglePIDLinearOp extends LinearOpMode {
         yawPIDController.setOutputRange(MIN_MOTOR_OUTPUT_VALUE, MAX_MOTOR_OUTPUT_VALUE);
         yawPIDController.setTolerance(navXPIDController.ToleranceType.ABSOLUTE, TOLERANCE_DEGREES);
         yawPIDController.setPID(YAW_PID_P, YAW_PID_I, YAW_PID_D);
+        yawPIDController.enable(true);
 
-        waitForStart();
-
-        while ( !calibration_complete ) {
+        while (!calibration_complete) {
             /* navX-Micro Calibration completes automatically ~15 seconds after it is
             powered on, as long as the device is still.  To handle the case where the
             navX-Micro has not been able to calibrate successfully, hold off using
@@ -118,55 +144,96 @@ public class ConceptNavXRotateToAnglePIDLinearOp extends LinearOpMode {
             calibration_complete = !navx_device.isCalibrating();
             if (!calibration_complete) {
                 telemetry.addData("navX-Micro", "Startup Calibration in Progress");
+                telemetry.update();
             }
         }
         navx_device.zeroYaw();
 
+        waitForStart();
 
-        try {
-            yawPIDController.enable(true);
+
 
         /* Wait for new Yaw PID output values, then update the motors
            with the new PID value with each new output value.
          */
 
-            final double TOTAL_RUN_TIME_SECONDS = 30.0;
+        final double TOTAL_RUN_TIME_SECONDS = 10.0;
+        int DEVICE_TIMEOUT_MS = 500;
+        navXPIDController.PIDResult yawPIDResult = new navXPIDController.PIDResult();
+
+        /* Drive straight forward at 1/2 of full drive speed */
+        double drive_speed = 0.5;
+
+        DecimalFormat df = new DecimalFormat("#.##");
+
+        rotateNew(90);
+    }
+
+
+    /**
+     * rotate to a certain angle
+     *
+     */
+    public void rotateNew(double targetAngle)  {
+
+        /* If possible, use encoders when driving, as it results in more */
+        /* predictable drive system response.                           */
+        //leftMotor.setChannelMode(DcMotorController.RunMode.RUN_USING_ENCODERS);
+        //rightMotor.setChannelMode(DcMotorController.RunMode.RUN_USING_ENCODERS);
+
+        /* Configure the PID controller */
+        yawPIDController.setSetpoint(targetAngle);
+        yawPIDController.setContinuous(true);
+        yawPIDController.setOutputRange(MIN_MOTOR_OUTPUT_VALUE, MAX_MOTOR_OUTPUT_VALUE);
+        yawPIDController.setTolerance(navXPIDController.ToleranceType.ABSOLUTE, TOLERANCE_DEGREES);
+        yawPIDController.setPID(YAW_PID_P, YAW_PID_I, YAW_PID_D);
+
+//        navx_device.zeroYaw();
+
+//            yawPIDController.enable(true);
+
+        /* Wait for new Yaw PID output values, then update the motors
+           with the new PID value with each new output value.
+         */
+
+        try {
+
+            final double TOTAL_RUN_TIME_SECONDS = 500.0;
             int DEVICE_TIMEOUT_MS = 500;
             navXPIDController.PIDResult yawPIDResult = new navXPIDController.PIDResult();
 
             DecimalFormat df = new DecimalFormat("#.##");
 
-            while ( (runtime.time() < TOTAL_RUN_TIME_SECONDS) &&
+            while ((runtime.time() < TOTAL_RUN_TIME_SECONDS) &&
                     !Thread.currentThread().isInterrupted()) {
                 if (yawPIDController.waitForNewUpdate(yawPIDResult, DEVICE_TIMEOUT_MS)) {
                     if (yawPIDResult.isOnTarget()) {
-                        leftMotor.setPowerFloat();
-                        rightMotor.setPowerFloat();
+                        frontleft.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
+                        frontright.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
+                        backleft.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
+                        backright.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
                         telemetry.addData("PIDOutput", df.format(0.00));
+                        telemetry.update();
                     } else {
                         double output = yawPIDResult.getOutput();
-                        leftMotor.setPower(output);
-                        rightMotor.setPower(-output);
+                        frontleft.setPower(output);
+                        backleft.setPower(output);
+                        frontright.setPower(-output);
+                        backright.setPower(-output);
                         telemetry.addData("PIDOutput", df.format(output) + ", " +
                                 df.format(-output));
+                        telemetry.update();
                     }
                 } else {
-			    /* A timeout occurred */
+                    /* A timeout occurred */
                     Log.w("navXRotateOp", "Yaw PID waitForNewUpdate() TIMEOUT.");
                 }
-                telemetry.addData("Yaw", df.format(navx_device.getYaw()));
+//                telemetry.addData("Yaw", df.format(navx_device.getYaw()));
+//                telemetry.update();
             }
         }
         catch(InterruptedException ex) {
-             Thread.currentThread().interrupt();
+            Thread.currentThread().interrupt();
         }
-        finally {
-            navx_device.close();
-            telemetry.addData("LinearOp", "Complete");
-        }
-
-
-
-
     }
 }
